requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Django==3.1.3
idna==2.10
joblib==1.0.1
nltk==3.6.2
numpy==1.19.4
opencv-contrib-python==4.5.1.48
opencv-python==4.4.0.46
pandas==1.1.5
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.24.0
scikit-learn==0.23.2
scipy==1.6.3
six==1.16.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.60.0
urllib3==1.25.11
